passed all tests with USE_CUDA
