+++ core/include/miso/boundary_condition.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <miso/boundary_condition_base.hpp>
+#include <miso/boundary_condition_core.hpp>
+#include <miso/boundary_condition_cpu.hpp>
+
+#ifdef USE_CUDA
+#include <miso/boundary_condition_gpu.cuh>
+#endif
