+++ core/include/miso/grid.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <miso/grid_cpu.hpp>
+
 #ifdef USE_CUDA
 #include <miso/grid_gpu.cuh>
 #endif
-
-#include <miso/grid_cpu.hpp>
