+++ core/include/miso/mhd.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <miso/mhd_cpu.hpp>
+
 #ifdef USE_CUDA
 #include <miso/mhd_gpu.cuh>
 #endif
-
-#include <miso/mhd_cpu.hpp>
