+++ core/tests/parallel/test_model_mpi.cu
@@ -2,7 +2,7 @@
 #include "test_model_mpi_common.hpp"
 
 TEST_CASE("Test Model GPU") {
-  miso::Model<Real> model = run_test_model();
+  auto model = run_test_model();
   REQUIRE(model.grid_d.i_total == model.grid_local.i_total);
   REQUIRE(model.grid_d.j_total == model.grid_local.j_total);
   REQUIRE(model.grid_d.k_total == model.grid_local.k_total);
